refactor: move commands to typescript
