+++ src/commands/price.ts
@@ -1,21 +1,25 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const alpha = require('alphavantage')({ key: process.env.ALPHAKEY });
 
-module.exports = {
+export = {
   data: new SlashCommandBuilder()
-    .setName('company-info')
-    .setDescription('Get company info of a ticker')
+    .setName('price')
+    .setDescription('Get price of a ticker')
     .addStringOption((option) => option.setName('ticker').setDescription('The ticker to get a quote')),
   async execute(interaction) {
     const ticker = interaction.options.getString('ticker');
     if (ticker) {
-      alpha.fundamental
-        .company_overview(ticker)
-        .then((data) =>
-          interaction.reply(
-            `Company overview of \`${ticker}\`: ${data.Name} -- ${data.AssetType} \n${data.Description}`
-          )
-        )
+      alpha.data
+        .quote(ticker)
+        .then((data) => {
+          console.log(data);
+          const price = data['Global Quote']['05. price'];
+          const percent = data['Global Quote']['10. change percent'];
+          const previousclose = data['Global Quote']['08. previous close'];
+          return interaction.reply(
+            `Current price of \`${ticker}\`: ${price} (${percent}) Previous close: ${previousclose}`
+          );
+        })
         .catch((error) => {
           console.error(error);
           return interaction.reply('Error has occurred');
